pyproject.toml
[project]
name = "supercat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "langchain-core>=0.3.78",
    "langchain-google-genai>=2.1.12",
    "langchain-tavily>=0.2.12",
    "langgraph>=0.6.9",
    "modal>=1.1.4",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "python-telegram-bot>=22.5",
    "pytubefix>=10.0.0",
    "uvicorn>=0.37.0",
    "yt-dlp>=2025.9.26",
]
